Splitted tests into several files. And some more on async
